Move fake link gateway to new module
